shelves ID fix from Integer to Long

+++ src/main/java/nl/marisabel/imReadingAPI/shelves/ShelvesRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShelvesRepository extends CrudRepository<ShelvesEntity, Integer> {
+public interface ShelvesRepository extends CrudRepository<ShelvesEntity, Long> {
 }
